Oncotree-package(Oncotree)	R Documentation

Constructing and evaluating oncogenetic trees

Description

Oncogenetic trees are directed tree structures that model the process of occurrence of genetic alterations during carcinogenesis.

Details

Package: 	 Oncotree

Type: 	 Package

Version: 	 1.0

Date: 	 2007-01-30

License: 	 GPL 2 or newer

A pure oncogenetic tree is a directed rooted tree T with a probability π (e) attached to each edge e such that for every vertex there is a unique directed path from the root to it along the edges of the tree. This tree generates observations on the presence/absence of genetic events the following way: each edge e is independently retained with probability π (e); the set of vertices that are still reachable from the root gives the set of the observed genetic evets.

To describe random deviations from the pure tree model an error model is added.

Error model

The tumor develops according to the pure oncogenetic tree model
The presence/absence of each alteration is independently measured
If the alteration is present it is not observed with probability ε_-.

If the alteration is absent it is observed with probability ε_+.
Author(s)

Lisa Pappas, Aniko Szabo

Maintainer: Aniko Szabo <aszabo@mcw.edu>

References

[1] Desper R., Jiang F., Kallioniemi O.P., Moch H., Papadimitriou C.H., and Sch"affer A.A. (1999) Inferring tree models for oncogenesis from comparative genome hybridization data. Journal of Computational Biology. 6m 37–51. [2] Szabo, A. and Boucher, K. (2002) Estimating an oncogenetic tree when false negative and positives are present. Mathematical Biosciences, 176/2, 219–236.

Examples


  data(ov.cgh)
  ov.tree <- oncotree.fit(ov.cgh)
  plot(ov.tree, edge.weights="estimated")

[Package Oncotree version 0.3 Index]