error.rates<-(Oncotree)	R Documentation

Set the error rates of an oncotree manually

Description

Allows to set the false positive and false negative error rate associated with an object of class oncotree to values other than those found by the optimization in oncotree.fit. The estimated edge transition probabilities are updated appropriately.

Usage


error.rates(x) <- value

Arguments

x 	 An object of class oncotree.

value 	 A numeric vector of length 2. The false positive error rate will be set to value[1], while the false negative error rate to value[2].

See Also

oncotree.fit

Examples


  data(ov.cgh)
  ov.tree <- oncotree.fit(ov.cgh)
  ov.tree
  error.rates(ov.tree) <- c(0,0)
  ov.tree

[Package Oncotree version 0.3.1 Index]