ov.cgh(Oncotree)	R Documentation

Ovarian cancer CGH data

Description

This is a data set obtained using the comparative genomic hybridization technique (CGH) on samples from papillary serous cystadenocarcinoma of the ovary. Only the seven most commonly occurring events are given.

Usage

data(ov.cgh)

Format

A data frame with 87 observations on the following 7 variables.

8q+

a 0/1 indicator of the presence of the ‘8q+’ event

3q+

a 0/1 indicator of the presence of the ‘3q+’ event

5q-

a 0/1 indicator of the presence of the ‘5q-’ event

4q-

a 0/1 indicator of the presence of the ‘4q-’ event

8p-

a 0/1 indicator of the presence of the ‘8p-’ event

1q+

a 0/1 indicator of the presence of the ‘1q+’ event

Xp-

a 0/1 indicator of the presence of the ‘Xp-’ event

Details

The CGH technique uses fluorescent staining to detect abnormal (increased or decreased) number of DNA copies. Often the results are reported as a gain or loss on a certain arm, without further distinction for specific regions. It is common to denote a change in DNA copy number on a specific chromosome arm by prefixing a “-” sign for decrease and a “+” for increase. Thus, say, -3q denotes abnormally low DNA copy number on the q arm of the 3rd chromosome.

Source

http://www.ncbi.nlm.nih.gov/sky/

Examples


  data(ov.cgh)
  heatmap(data.matrix(ov.cgh), Colv=NA, scale="none", col=c("gray90","red"))

[Package Oncotree version 0.3 Index]